refactor keyword utils
